refactor: unify event emission in AbstractClient
- Replaced individual event emission methods with a single `emit` method in `AbstractClient` to streamline the handling of request lifecycle events.
- Introduced `HTTPClientEvent` interface and corresponding event classes for better structure and consistency.
- Updated event handling in `RequestStarted`, `RequestSucceeded`, and `RequestFailed` classes to utilize the new event system.
- Added unit tests for the new event classes to ensure proper functionality.
